+++ src/commands/thermal_resistances/thermal_resistance_75.rs
@@ -27,7 +27,7 @@ pub struct ThermalResistance75 {
 impl ThermalResistance75 {
     /// Returns the thermal resistance value in ohms at 25 degrees Celsius.
     pub fn thermal_resistance(&self) -> ElectricalResistance {
-        ElectricalResistance::new::<ohm>(u32::from(self.raw_thermal_resistance()))
+        ElectricalResistance::new::<ohm>(self.raw_thermal_resistance())
     }
 }
 impl_two_byte_read_command!(ThermalResistance75, Command::ThermalResistance75);
